Always send the default icon in a callout block
